requirements.txt
keras==3.13.2
numpy==2.2.6
pandas==3.0.1
opencv-python==4.12.0.88
matplotlib==3.10.8
scikit-learn==1.8.0
scipy==1.17.1
pillow==12.1.1
jupyter==1.1.1
jax==0.4.38
jaxlib==0.4.38
opt_einsum==3.4.0
